fix: rename sse to mevshare
